Fix audit producer test (#284)
- Take in account new events thrown for signal received event that was not thrown before
- Move test to SignalAuditProducerIT

Related to https://github.com/Activiti/Activiti/issues/2272
